Validation added

+++ controllers/user.controller.js
@@ -4,9 +4,15 @@ const { message } = require("../constants/messages.constants");
 const { responseStructure: rs } = require("../helpers/response.helper");
 const { updateFilters, FiltersMeta } = require("../helpers/filter.helpers");
 const { getSelectString, SelectMeta } = require("../helpers/dbselect.helper");
-const { getCleanObject } = require("../helpers/index.helper");
+const { getCleanObject, getPayload } = require("../helpers/index.helper");
 const { jwt_key } = require("../config/env.config");
 const jwt = require("jsonwebtoken");
+const {
+  isRequestBodyForAddRecordValid,
+  isRequestBodyForUpdateRecordValid,
+  getObjectWithValidFields,
+} = require("../helpers/validation.helper");
+const { UserSchema } = require("../models/user.model");
 
 /**
  * @swagger
@@ -140,9 +146,19 @@ const jwt = require("jsonwebtoken");
 // Add user
 exports.addUser = async (req, res) => {
   const jsonData = req.body;
-  //Validation
+
+  const { isValid, missingFields, validFields } =
+    isRequestBodyForAddRecordValid(jsonData, UserSchema.schema);
+  if (!isValid) {
+    return res
+      .status(status.badRequest)
+      .send(rs(status.badRequest, message.missingFields, { missingFields }));
+  }
+
+  const validObject = getObjectWithValidFields(jsonData, validFields);
+
   userdb
-    .addUser(jsonData, jsonData.email)
+    .addUser(validObject, jsonData.email)
     .then((response) => {
       res
         .status(status.success)
@@ -257,6 +273,13 @@ exports.getUsers = (req, res) => {
 // Get single user data
 exports.singleUser = (req, res) => {
   let userId = req.params.id;
+
+  if (!userId) {
+    return res
+      .status(status.badRequest)
+      .send(rs(status.badRequest, message.noUniqueId));
+  }
+
   const selectString = getSelectString(SelectMeta.default, SelectMeta.users);
   userdb
     .getUsers({ _id: userId }, { select: selectString })
@@ -321,6 +344,11 @@ exports.singleUser = (req, res) => {
 // delete user
 exports.deleteUser = (req, res) => {
   let userId = req.params.id;
+  if (!userId) {
+    return res
+      .status(status.badRequest)
+      .send(rs(status.badRequest, message.noUniqueId));
+  }
   userdb
     .deleteUser(userId)
     .then((response) => {
@@ -384,11 +412,24 @@ exports.deleteUser = (req, res) => {
 
 // update user
 exports.updateUser = async (req, res) => {
-  const userId = req.body._id;
+  const decoded = getPayload(req);
   const jsonData = req.body;
+  const userId = jsonData._id;
+  const { isValid, validFields } = isRequestBodyForUpdateRecordValid(
+    jsonData,
+    UserSchema.schema
+  );
+
+  if (!isValid) {
+    return res
+      .status(status.badRequest)
+      .send(rs(status.badRequest, message.noUpdateFields));
+  }
+
+  const validObject = getObjectWithValidFields(jsonData, validFields);
 
   userdb
-    .updateUser(userId, jsonData, systemUser)
+    .updateUser(userId, validObject, decoded.email)
     .then((response) => {
       res.status(status.success).send(rs(status.success, message.userUpdate));
     })
@@ -444,16 +485,17 @@ exports.updateUser = async (req, res) => {
 
 // Login user
 exports.login = async (req, res) => {
-  const jsonData = req.body;
-  //Validation
+  const { email, password } = req.body;
+  if (!email && !password) {
+    return res
+      .status(status.success)
+      .send(rs(status.unauthorized, message.unauthorized));
+  }
   try {
-    const users = await userdb.getUsers(
-      { email: jsonData.email },
-      { lean: true }
-    );
+    const users = await userdb.getUsers({ email: email }, { lean: true });
     if (users.length > 0) {
       const user = users[0];
-      if (user.password === jsonData.password) {
+      if (user.password === password) {
         const accessToken = jwt.sign(
           getCleanObject(user, SelectMeta.default, SelectMeta.users),
           jwt_key,
@@ -534,7 +576,7 @@ exports.refreshToken = async (req, res) => {
     const parts = tokenString.split(" ");
     const token = parts[1];
     if (!token) {
-      res
+      return res
         .status(status.success)
         .send(rs(status.unauthorized, message.unauthorized));
     }
